requirements.txt
openai>=1.14.0
python-dotenv>=1.0.1
requests>=2.31.0
rich>=13.7.0
